feat: initial work to migrate to a console project instead of tests project
## Reasoning

Tests project meant that if a single entity for a type failed validation, then the rest of the checks would be skipped for that type.

By migrating it to a console project, we can simply run all the checks and log messages as needed.

It's a bit of extra work, but this commit lays the groundwork for the majority of it.