fix: Reduce page query
The existing page query resulted in a **massive** SQL query which often timed out, especially on more complex pages.

Whilst this has been fixed in various other places (e.g. the web app page query), it was missed from the Azure Function.

Instead of querying _everything_ for the page, it now only returns the necessary fields for the page, as well as a _collection of IDs_ for the references. This ensures the query takes around 50ms at most.