fix(e2e): :adhesive_bandage: Add videos on E2E failures again (#622)
* fix(e2etests): :bug: Record videos on Cypress failures again

This functionality was disabled by default in Cypress 13. So when we upgraded the packages recently it disabled this functionality.

This fix adds video recording for failures again.

* chore(cicd): :adhesive_bandage: Ignore "missing" files on E2E tests

If no screenshots or videos are found for artifact upload, by default it is treated as warning.

This change ignores them instead, as we should not have anything on successes.