Make non-nullable properties `required` within `RecommendationsViewModel`.
Reviewing usages of these non-nullable properties, I see that there are zero nullability checks around any of them and it appears to be appropriate to enforce there being a value provided. For example, the `CategoryName` is displayed within an `<h1>` and the `SectionName` is frequently called with `ToLower()` (where a null value would result in an NRE).

Note that there is one usage where I wasn't able to provide a `CategoryName`. In this case it seems to be a route by `maturity`, which is now not used within the web UI for PlanTech (data remains in the database). Given this, I have opted to remove the now-unused controller method instead of trying to adapt/modify it.